feat:  разбиение js-кода на модули
