+++ js/data.js
@@ -1,0 +1,55 @@
+
+function getDataForPhotos() {
+
+  const DESCRIPTIONS = [
+    'Самое красивое место',
+    'Солнечное утро',
+    'Необычный закат',
+    'Тихая гладь озера',
+    'Осенний лес',
+    'Беззаботная пора',
+    'Сама милота',
+    'Самое красивое место',
+    'Солнечное утро',
+    'Необычный закат',
+    'Тихая гладь озера',
+    'Осенний лес',
+    'Беззаботная пора',
+    'Сама милота',
+    'Самое красивое место',
+    'Солнечное утро',
+    'Необычный закат',
+    'Тихая гладь озера',
+    'Осенний лес',
+    'Беззаботная пора',
+    'Сама милота',
+    'Самое красивое место',
+    'Солнечное утро',
+    'Необычный закат',
+    'Тихая гладь озера',
+    'Волшебный закат'
+  ];
+
+  const COMMENTS = [
+    'Всё отлично!',
+    'В целом всё неплохо. Но не всё.',
+    'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+    'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+    'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+    'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
+  ];
+
+  const NAMES = [
+    'Иван',
+    'Татьяна',
+    'Мария',
+    'Владимир',
+    'Виктор',
+    'Юлия',
+    'Петр',
+    'Сергей',
+  ];
+  return {DESCRIPTIONS, COMMENTS, NAMES};
+}
+
+export {getDataForPhotos};
