+++ js/utils.js
@@ -1,0 +1,16 @@
+const getRandomInteger = (a, b) => {
+  const lower = Math.ceil(Math.min(a, b));
+  const upper = Math.floor(Math.max(a, b));
+  const result = Math.random() * (upper - lower + 1) + lower;
+  return Math.floor(result);
+};
+
+function getIndexIncrement () {
+  let count = 0;
+  return function () {
+    return ++count;
+  };
+}
+
+export {getRandomInteger, getIndexIncrement};
+
